refactor: collapse identical prose generators into shared default
The four host-specific prose files (claude-code, codex, opencode,
generic) were byte-for-byte identical. Replaced with a single
default.ts and an override registry in index.ts. To add host-specific
prose, spread defaultGenerator and replace individual methods.
